+++ testing/mainTest.py
@@ -1,0 +1,9 @@
+from importFile import booba
+
+x = 20
+
+print(f"x is now {x}")
+
+x = booba(x)
+
+print(f"x is now {x}")
